hopr_transport_p2p
Module multiaddrs

Utilities for work with multiaddresses

Structs

Multiaddr

Representation of a Multiaddr.

Enums

Protocol

Protocol describes all possible multiaddress protocols.

Functions

is_dns

Check whether the first multiaddress protocol component is a dns* component

is_private

Check whether the first multiaddress protocol component represents a private address

is_supported

Check whether the multiaddr protocol component is supported by this library

strip_p2p_protocol

Remove the p2p/<PeerId> component from a multiaddress