hopr_transport_p2p

Module swarm

Raw swarm definition for the HOPR network.

Structs

HoprSwarm

HoprSwarmWithProcessors

Enums

Inputs

Composition of all inputs allowing to produce a single stream of input events passed into the swarm processing logic.

Type Aliases

TicketAggregationEvent

TicketAggregationRequestType

TicketAggregationResponseType

TicketAggregationWriter